Lazy dependencies for factories from the factory provider.
TyphoonAssistedFactoryBase uses the new shouldInjectProperty:withType:lazyValue:
to store the lazy value block in its internal map, and telling Typhoon to not
inject the property.

The creation process no longer creates the setter (not needed anymore), but
creates an ivar for each property (same name as the property).

Later, when the getter of the property is invoked, if the ivar is still nil,
the lazy value block is invoked, saved into the ivar and returned to the
caller. The properties are synchronized to avoid two threads creating the ivar
at the same time.

See #145